[core] Use cpp files for shader source
